😬 Extensions are bad bad
