+++ lib/televerse.dart
@@ -38,8 +38,6 @@ export 'src/types/types.dart';
 // Bot & API
 part 'src/televerse/api/raw_api.dart';
 part 'src/televerse/bot/bot.dart';
-part 'src/televerse/bot/listeners.dart';
-part 'src/televerse/bot/extras.dart';
 
 // Context
 part 'src/televerse/context/constructor.dart';
@@ -49,7 +47,6 @@ part 'src/televerse/context/properties.dart';
 
 // The On Method
 part 'src/televerse/filters/filters.dart';
-part 'src/televerse/filters/on.dart';
 
 // Keyboards & Menus
 part 'src/televerse/markups/inline_keyboard.dart';
